cube-network: disable network-manager on veth* pair interfaces
While we should rely on an image to ignore interfaces that are created
and managed by cube-network .. it is not always a possible for those
images/containers to know the name of the interface.

While creating a network device pair, we do know the interface and we
can detect NetworkManager in a container, and tell it to ignore our
newly created interface.

We could also do this for passthrough devices, but that is something
that a user may want to manage with NetworkManager, so we leave it
alone for now.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>