cube-network: Fix regressions from patches that were dropped
commit ca060b46241251d6c5d97cdc571a116569b57d48 (overc-conf: drop
oci-network, since cube-network is now universal) accidentally dropped
two patches that were used in the oci-network.  The older patches
cleanly apply so long as the veth-br-int is changed to ${veth_name}.

The MAC address is changing after dhclient is running which causes a
problem where you lose IP connectivity after a while when the lease
renews.  It also poses a problem where you cannot get a lease in a
timely manner because the local lease db was being shared across
containers, so for the first container everything was fine, but each
additional container could incur a delay.

--- Restore commit ca060b46241251d6c5d97cdc571a116569b57d48 ---

oci-network: Use a private lease file for each container for dhclient

The dhclient state file cannot be shared if the interface name is the
same in a container.  It will cause the wrong IP address to be passed
for a renew request.

--- Restore commit feb6b8ef13fb6ebeb6e476defaa31d93cc3f4872 ---

oci-network: Move MAC assignment before dhclient and use a hash function

The MAC assignment needs to come before the dhclient request else the
network address changes later on when it renews the lease.

For speed and consistent operations the MAC address should be the same
across container restarts so we can use a hash function to generate
the mac address.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>