dhcp,dhclient: Add --psig option to die when the parent process exits
When using dhclient in a hook function a runc monitor process waits to
execute the hooks and clean up.  As a part of the clean up when this
function exits any dhclient processes should also die so that all the
network name spaces are released immediately.

This patch adds the --psig to send a SIGKILL when the parent process
terminates.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>