cube-ctl: Add a randomized uuid for any created container
The latest version of systemd expects that a container will have the
variable container_uuid controlled by the container manager.  If an
end user does not set the container_uuid with c3 cfg set -n CONTAINER
cube.env:container_uuid the c3 add command will have done it for them.

Failure to set the container_uuid in a network prime container results
in programs like the Network Manager using a new client ID for each
reboot, which in turn causes you to get a new IP address on each
reboot.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>