cube-network: add c3-ipcfg to abstract dynamic networking details
Introducing a utility c3-ipcfg to abstract the details of how we
get dynamic IP addresses.

As part of this change, we make the default "dynamic" network be
a call to c3-ipcfg, and move our dhclient based configuration to
"dynamic-dhcp". This will allow us to fall back to the proven
methods if the new implementation has details.

Note: c3-ipcfg is not completed yet, since it doesn't wrap dhcp
or etcd based allocations. That is future work, and will be
completed as required.

The current/default scheme for c3-ipcfg is to use mac address
based dynamic IP assignment. This is a hashing of the mac
address (which is already unique, and derived from the machine ID)
to an ip address in our internal range.

There shouldn't be any collisions, and they are not dealt with
at the moment, since there is no global tracking of IPs.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>