cube-vrf: Introduce the cube-vrf container
Move the OpenVSwitch 'br-int' bridge from -essential to a new
container (cube-vrf) along with updates to the network hooks to push
the container veth ends that would have previously existed in
-essential into the new cube-vrf container.

These changes go along with changes to the install scripts which
establish cube-vrf as a pflask container which will be run at startup
via the essential-autostart.service. Since 'c' in 'cube-vrf' comes
before 'd' in 'dom0' the cube-vrf container will be the first to start
at bootup. This ensures that 'dom0' veth interface will be able to be
moved into the cube-vrf container.

With these changes applied the system will look much as it did before,
with the following changes:
 * br-int in -essential will only have the br-int port and IP
 * the cube-vrf will be listed in machinectl and c3
 * the cube-vrf hosts a new br-int OVS bridge where all
   container veth interfaces are now attached

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>