cube-essential: remove all traces of openvswitch
With the introduction of the cube-vrf container we have moved the main
cube lan bridge out of essential and the br-int which remains in
essential is now unused. Remove it and all configs and such which
reference it. This will be replaced by a veth pair which will connect
essential to the cube lan via the cube-vrf.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>