cube-essential: setup networking connecting to cube-vrf
In cube-essential we adjust the networking some to drop the use of
ansible, opting instead to use ROOTFS_POSTINSTALL_COMMAND. The network
configuration and resolv.conf remain as they were previously
(ip=192.168.42.2/24, nameserver=192.168.42.1...). We add a 'netdev'
file which is used by systemd-networkd to create an 'essential' veth
pair. The veth peer can be used by cube-vrf or equivalent or could be
left unused.

The 'nslinkup' and 'nslinkdown' actions in the network hook will now
detect if they are being called in conjunction with cube-vrf and now
handle the "up'ing" of the interface and moving the peer into the
'cube-vrf' namespace.

Lastly the cube-vrf gets a permanent 'essential' port added to the
br-int. This saves having to create/destroy it along with the timing
issues that brings.

With this commit the behavior of the system is restored to how it was
prior to the introduction of the cube-vrf, except the 'br-int' now
exists outside of the root namespace.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>