overc-conftools: remove timezone setup from ansible playbook
In order to allow for cube-essential to be used in an initramfs it is
desireable to drop ansible from the cube-essential image. Dropping
these timezone items form the playbook is another step towards this
goal. This is also reflective of our current aims for cube-essential,
in that it should not be interacted with and as such always using GMT
as the timezone and not having it configurable is fine.

At this point we no longer include the timezone package (tzdata) in
the essential image so these playbook items are currently unused so
this change does not affect the runtime in any way.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>