overc-conftools: remove -essential initial_user and password setup
In our effort to allow cube-essential to be used as an initramfs we
are preparing to remove ansible from cube-essential. The last items
left in essential.yml are there setup an 'initial_user' and password
(for that user). At one point having a user other than 'root' on
-essential made sense but this is no longer the case so we can savely
remove this functionality.

Since this leaves the essential.yml playbook/file empty we can drop it
and any remaining references to it.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>