overc-conftools: be more specific finding cube_vrf_pid
When cube-vrf is starting up there are several PIDs which match our
search criteria for both the PPID and the container PID. So although
the search works fine in many cases, and always works after startup,
it is often failing, resulting in networking not being in the UP
state.

For the PPID we can assume that init/systemd is always going to be its
PPID, so we can add the PPID==1 filter to narrow down the search. When
using the PPID to find the container PID we can use the fact that the
container process will be running '/sbin/init', both during startup
and after the system settles (though the process string is not quite
the same the filter works in both scenarios). With these two filters
added to the mix we get the right PPID and PID and the networking for
dom0 always ends up in the UP state.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>