vrf: make VRF container flexible and move DNS to the VRF
The VRF should not only take care of the vSwitch, it should also
take care of the DNS services to simplify dom0 (which was hardcoded
as the VRF before).

With this change, we move dnsmasq into the VRF container and change
the init of the VRF to start dnsmasq, but not OVS.

To make the vrf container flexible, we must make the VRF addressable
(at 192.168.42.4) and assign that IP to the br-int interface of the
VRF.

We move the setup of the OVS instance into the cube-network hook,
since this is the best and simplest way to ensure that the vswitch
is up, br-int is available and other pre-conditions are met to
assign 192.168.42.4.

Finally, we change the hooks to forward DNS traffic to .4 (versus
dom0 at .3).

Now that the VRF is flexible, we also tweak the hooks to lookup
the vrf container by its attribute, versus any fixed name.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>