overc-conftools: move system-id script out of the ansible playbook
The last thing left in the ansible playbook is a call to
'systemid-set.sh', to setup the '/etc/system-id' file used by dnf. We
still need to execute this script so here we provide a "low tech"
replacement for ansible before dropping it completely from the image.

To create the "low tech" ansible replacement we create a new script
called 'overc-conf.sh' which will execute any scripts found in
'/etc/overc-conf/overc-conf.d'. We set 'overc-conftools.service' to
execute 'overc-conf.sh' instead of running through the ansible
playbook. Lastly we move 'systemid-set.sh' from its current location
into '/etc/overc-conf/overc-conf.d' such that it is executed by our
ansible replacement.

If we have any other runtime configuration changes to add in the
future we will be able to place them in the
'/etc/overc-conf/overc-conf.d' directory and they will simply be
executed.

At this point we are not longer using any ansible in the essential
image and it can be removed.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>