vrf/dns: fix DNS startup race conditions
The vrf must be started before dom0 or the netprime, which must register
with the OVS instance that runs within the VRF.

To do that, we implement the ability to process a simple 'depends' file
in a container directory. That file has the name of a container that must
be started before the named one.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>