cube-network: document 'self' networking mode
Enable the case condition for a 'self' networking mode where the veth
pair is not assigned an address. It is up to the container to manage
the interface in this scenario.

There are plenty of ways to break things with this option, but we allow
it for some indpendent containers.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>