dnsmasq: update bbappend to work with recent dnsmasq version bump
The dnsmasq recipe was recently bumped to v2.79 which resulted in our
bbappend no longer being used. Use a wildcard to allow the bbappend to
be applied. We should most likely drop this bbappend soon as how we
use dnsmasq has changed with the introduction of the VRF making our
local changes no longer necessary.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>