linux-yocto: allow linux-yocto and linux-yocto-dev to have distinct patches
We have no plans to push the 'uncontain' patches to the Yocto kernel
or upstream and thus plan to host them in this layer for the
foreseeable future. As such the kernel versions in play for linux-yocto
and linux-yocto-dev can significantly diverge making the patches fail
to apply to both kernels. Allow for different versions of these
patches to exist to allow for us to continue to successfully patch and
build both kernels.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>