meta-cube/layer.conf: Adjust distro configurations out of cube layer.conf
The Yocto Project compatible requirements require that a layer can't just
change distribution wide settings without being specifically enabled.

Adjust the system to only make these settings if the overc distribution is
enabled.  It must be done in this order, as bitbake.conf (which loads
local.conf) is loaded AFTER the layers.conf files.  But the class configs
are then loaded after the bitbake.conf, so we can forceably load our own
configurations after bitbake.conf if certain conditions are met.

In this case, we specifically check if the distribution is set to 'overc',
if it is -- we then add the configurations specific to overc.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>