meta-cube/layer.conf: Only .bb and .bbappends if we're processing the overc distro
Yocto Project Compatible requires that a layer may not modify (bbappend)
components without that modification being specifically enabled by the user.

For instance, distribution specific changes should not be enabled, unless
that distribution is enabled.

In this case, both meta-overc and meta-overc/meta-cube are only valid when
used with the overc distribution.  So adjust the BBFILES to also enfore
this, and thus permitting the layers to pass the Yocto Project Compatible
test for signature differences.

We also do not want the .bb to be loadable without overc being enabled.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>