housekeeping: add cube-vrf to docs and scripts
We have added a new container to the OverC framework to handle
networking services for the cube-lan. Although this functionality has
been available since commit ad749f0c0660 [cube-vrf: Introduce the
cube-vrf container], we have yet to include it in documents and system
scripts. Add references to cube-vrf where needed.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>