cube-netconfig: Move dhclient for netprime to netprime pid space
We want to make sure that anytime the netprime container dies
ungracefully, or is killed intententionally that all its associated
processes are also reaped automatically.

This solves the problem where the netprime container was restarted and
the stop hook did not run which would have properly cleaned up the
stray dhclient process.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>