cgmanager: delete recipe, upstream has been deprecated
As documented in the README here
https://github.com/lxc/cgmanager

and details here:
https://s3hh.wordpress.com/2016/06/18/whither-cgmanager

this package has been incompatible with versions of systemd since v232
and has been deprecated upstream.

Along with the fact that we have not included cgmanager in of the
images which are part of OverC for a while, we are going ahead and
removing this recipe.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>