cube-vrf: Use the tini 'init'
Currently we use a simple bash script as our init. Although this does
the job it lacks support for reaping zombies and proper signal
handling support. It is common to use a basic 'init' such as tini or
dumb-init in containers to get this functionality. Our script is still
used, called by tini, but tini now becomes our init.

In order to use this change you must have the change to the
overc-installer which allows for the process.app to be defined.

NOTE: we still 'sleep inifinity' since tini exits when the program it
calls exits.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>