cube-cfg: don't insert newline when there are multiple process args
Currently when jq is used to parse the process.args to produce the
pflask.cmd file it outputs one arg per line. So for example if you
have something like this in the config.json:

  "process": {
    "args": [
      "/sbin/init",
      "/blah"
    ],
  }

the generated pflask.cmd will be something like:

  pflask ... -- /sbin/init
  /blah

which isn't what is expect and can arguable be seen as
useless. Normally when multiple process.args are passed they are
expected to be in the form:

  pflask ... -- <program> <arg> ...

as is the case with process.args when used with an OCI container. With
this change we change the output of jq to print the args in a single
line with a space delimiter. This will result in a pflask.cmd like:

  pflask  --mount ... -- /sbin/init /blah

Making the pflask and OCI handling of multiple process.args match.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>