cube-vrf: Use /etc/vrf-resolv.conf to establish dns forwarding dynamically
There is a race condiditon from the time dnsmasq is terminated to the
time it can be restarted and bind to its ports and start forwarding
DNS queries.

We can solve the problem by using an alternate file such as
/etc/vrf-resolv.conf to take the place of resolv.conf.  The dnsmasq
server will use an inotify event to trigger a reload of the dns
fowarding information to resolve queries any time the file is
rewritten.  This means we no longer need to restart the process or
pass a signal.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>