tini: don't attempt to place link for /sbin/init
Originally I thought it would be a good idea to have a link to tini
'init' in /sbin/init, to keep things where people expect them. If
someone asked 'I wonder what init is being used?' it would be easy to
find out. Unfortunately tini is also included in other images where it
is present but not used as the 'init', in this case /sbin/init was
conflicting. So although this was a good idea it fell over in
practice, drop the /sbin/init link, we will use tini in its default
install location.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>