python-ansible: remove recipe
As we continued to develop the OverC framework we realized that our
use of ansible was interfering with some of our goals (such as image
size and boot times). We therefor dropped the use of ansible in favor
of using bash scripts in commit 03f3e1010715 [cube-essential: remove
last bits of ansible]. With ansible no longer in use we are now
removing its recipe here (moving it to meta-cloud-services).

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>