housekeeping: drop unused GO recipes
Since we removed consul and consul-migrate these recipes are no longer
in use and can be removed. These were only ever used a "source"
dependencies and so their output was only used to build other GO
packages and not deployed as part of any images, as such we lose
nothing in dropping them.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>