cube-essential: Fix the read-only root fstab
The /opt/container directory is a physical mount point for the volume
that contains the the containers, and should not be tmpfs.  The
openvswitch mount is no longer used in the root name space after the
uprev and reconfiguration of the openvswitch.  And the /var/lib/cube
tmpfs is handled differently by the essential startup script.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>