overc-conftools: Add read-only rootfs support
Nothing should write to the /etc directory, so the dhclient scripts
needed to be moved to write into their respective containers as well
as the per container resolv.conf files.

Also the master lease database cannot be shared when using multiple
containers, it needed to be moved as well into the correct container
space.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>