dom0-init,essential-init: Change /lib/firmware from a copy to a link
The current mechanism for handling /lib/firmware was a copy and sync
from dom0 -> essential using rsync and required a read-write file
system.

The need to support a fully read-only rootfs necessitated moving the
firmware copy to a different file system.  At that point system
/lib/firmware directory became unavailable at boot because it did not
live in / or /usr.  This was fixed with a new systemd rule to mount
/opt/container just after the root file system check is completed and
transitioned to either RO or RW.

Because the firmware had been moved to the same volume where it was
copied from, the copy was scheme was removed.  This allows for a
reduction in file system space, faster boot time after the firmware
has been updated or at a 1st boot, as well as simplified management of
the firmware.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>