dom0-init,essential-init: Change /lib/firmware from a copy to a link
The current mechanism for handling /lib/firmware was always a copy and
sync from dom0 -> essential.  This is problematic for the case of a
read-only root file system or a top level volume which uses slow media
backed with nonvolatile storage.

The read-only rootfs support added the ability to mount the
/opt/container volume early in the boot cycle which allows us to make
the optimization that we can have just a single copy of the firmware
in the dom0 container and avoid the complexity of syncing.  This is in
the new default.  Should a particular use case warrant syncing the
firmware to the /var/lib/cube/essential/lib/firmware directory all
that is needed is to break the link.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>