cube-cfg: enable device whitelisting
Enable the processing of device whitelisting in the oci-device hook.
When devices are added to the whitelist (via cube-cfg device whitelist:<device or pattern>
the hook will create the same nodes in a container during startup.

This allows us to not bind mount the entire /dev/, and be more selective

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>