cube-netconfig: allow a static IP on the netprime
In a scenario where a dhcp server is not available (or configured), it is
desireable to be able to set a static IP on the netprime's external
interface.

To do this, we add the processing of two variables to the netconfig hook:

  - cube.network.external.ip
  - cube.network.external.nameserver

An example of how to use them from dom0:

 % cd /opt/container/dom0
 % echo "128.224.56.218/24" > cube.network.external.ip
 % echo "search wrs.com" > cube.network.external.nameserver
 % echo "nameserver 147.11.57.128" >> cube.network.external.nameserver
 % echo "nameserver 149.11.57.133" >> cube.network.external.nameserver

Reboot, and the values will be used. This will eventually be exposed via
cube-cfg if it becomes a common enough operation.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>