skopeo: drop dependencies on other go source packages
Having worked with GO packages for a while now we have determined that
it is better to deal with GO dependencies using a SRC_URI approach
instead of with additional GO recipes and the DEPENDS value. This is
primarily due to the fact that every top level GO package will depend
on specific revisions of GO source dependencies and these versions
will almost always vary from top level GO package to the next. So it
really only makes sense to have recipes for top level GO packages,
such as skopeo. For this reason we have cleaned up the DEPENDS to
remove GO dependencies.

In the case of skopeo we do not need to replace these with SRC_URI
entries since skopeo's repository includes a fully populated 'vendor'
directory. Thus we can simply drop the DEPENDS and replace them with
nothing, as we are doing here. The end result should actually be the
same as the contents of the 'vendor' directory would always be
prefered over other sources of the dependent GO sources. So even
though we were making the GO dependencies available via DEPENDS
previously they were not being used (according to the GO documentation
of the 'vendor' directory)

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>