packagegroup-util-linux: add util-linux-nsenter
Work on commit 474d9329bea2 [overc-utils: RDEPEND on util-linux-X] was
imcomplete. We also need to ensure util-linux-nsenter is part of
cube-essential so we add it to the packagegroup-util-linux.

This fixes errors seen when running 'c3 list' in -dom0, which is
caused since cube-cmd will fail to execute nsenter in -essential.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>