linux-yocto: move to version 5.0
The version of linux-yocto has moved to version 5.0 and we must now
follow suit. Adjust all 4.18 references to 5.0.

We are able to drop the 4.18 specific versions of patches as the
patches apply cleanly to 4.19 and newer versions of the kernel. The
logic is left but commented out in case there are occassions for
version specific patches in the future.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>