systemd: update to oe-core 242 version
Updating the systemd bbappend to match the version in oe-core. We also
refresh patch 0001-systemd-make-udev-create-or-delete-device-nodes.patch
to apply to this systemd version.

But due to changes in systemd it no longer compiles. Since the utility
of udev changes is not currently required, we drop the patch from the
SRC_URI, but keep it around for the future.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>