README.overc-from-scratch.txt: cleanup and updates
There were several changes such as requiring the meta-selinux and
meta-cloud-services layers as well as requiring the use of cube-vrf
which made the current instructions out-of-date. These changes bring
the instructions back to a state where they can be cut and paste and
result in a working image.

Other changes to make this file useable include:

* adding the rng device to qemu (since 4.16 this is pretty much a must
  or the VM will 'stall' during boot)

* making the local.conf and config-live-dom0-server.sh use heredocs to
  make them more cut and paste friendly

* use of OVERCDIR to make directory selection more portable

* use of the 'master' branch for all repos

* specifying the GETTY otherwise the serial console is useless

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>