overc-system-agent: btrfs: don't fail if the kernel file isn't found
Currently if the kernel file isn't found we end up getting a python
backtrace when the md5sum is attempted to be calculated in
__init__. Obviously this isn't very user friendly so we have a choice:
error out or continue in the absence of a kernel file.

Since many operations do not involve the kernel file we have opted
here to not simply error out, but instead, if the kernel file is not
found we will skip operations involving the kernel and log this
behavior so that it is clear to the user that this is the case.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>