housekeeping: fixup systemd and python-jina2 dangling bbappends
Recipes for systemd and python(3)-jinja2 have been uprev'd so we need
to match these changes. For python-jinja2 we also move to the python3
version as python2 is EOL'd and has been removed.

The systemd "block devs" patch had to be updated as the files has
moved location in the upstream package, the patch itself remains the
same.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>