packagegroup-builder: cleanup py2 packages
With the removal of python2, due to it being EOLed, we need to make
some adjustments. Switch to python3- where it makes sense, drop ones
which would be duplicates, and in cases like python-compiler, the
package had been moved into core libs and eventually dropped in
python3, so drop it.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>